# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ePKL Backend API",
        "description": "Backend for ePKL mobile app",
        "contact": {
            "name": "Muhamamd Haikal",
            "email": "realityinaship@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/tempat-pkl": {
            "get": {
                "tags": [
                    "Tempat PKL"
                ],
                "summary": "Get all Tempat PKL",
                "description": "Retrieve all Tempat PKL records.",
                "operationId": "getAllTempatPkl",
                "parameters": [
                    {
                        "name": "users_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tempat PKL retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tempat PKL"
                ],
                "summary": "Create a Tempat PKL",
                "description": "Create a new Tempat PKL record.",
                "operationId": "createTempatPkl",
                "requestBody": {
                    "description": "Tempat PKL details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "users_id": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "type": "string"
                                    },
                                    "nama_tempat": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tempat PKL created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tempat-pkl/{id}": {
            "get": {
                "tags": [
                    "Tempat PKL"
                ],
                "summary": "Get a Tempat PKL by ID",
                "description": "Retrieve a specific Tempat PKL record by ID.",
                "operationId": "getTempatPkl",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tempat PKL",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tempat PKL retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tempat PKL"
                ],
                "summary": "Update a Tempat PKL",
                "description": "Update an existing Tempat PKL record.",
                "operationId": "updateTempatPkl",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tempat PKL",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Tempat PKL details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "users_id": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "type": "string"
                                    },
                                    "nama_tempat": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Tempat PKL updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tempat PKL"
                ],
                "summary": "Delete a Tempat PKL",
                "description": "Delete a Tempat PKL record by ID.",
                "operationId": "deleteTempatPkl",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tempat PKL",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tempat PKL deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}